+++ PolygonChecker/rectangleSolver.h
@@ -3,7 +3,7 @@
 #include <stdio.h>
 #include <stdbool.h>
 
-float calculateDistance(int x1, int y1, int x2, int y2);
+float calculateDistance(float x1, float y1, float x2, float y2);
 bool formsRectangle(int x[], int y[]);
 float rectanglePerimeter(float length, float width);
 float rectangleArea(float length, float width);
